refactor(media): move VlcStatus/parse_status_xml/newest_file to mur-common (pure movement)
The runtime's WatchScheduler can't depend on mur-core (Pillar-B), so it carried a
duplicate `newest_file` and parsed VLC state by raw string match. Move the shared,
logic-light pieces to `mur-common::media` (which both crates already depend on and
which already holds load_watch/save_watch/load_runtime):

- VlcStatus + parse_status_xml (adds quick-xml to mur-common)
- newest_file + newest_file_excluding

mur-core re-exports them from vlc.rs/scene.rs (no behavior change); the runtime
drops its private newest_file copy and uses the shared one. Tests move with the
code. No behavior change — the runtime's status check is swapped in a follow-up.
