+++ mur-common/Cargo.toml
@@ -16,6 +16,7 @@ serde_bytes = "0.11"
 serde_json = { workspace = true }
 serde_yaml = { workspace = true }
 serde_yaml_ng = "0.10"
+quick-xml = { workspace = true }
 chrono = { workspace = true }
 uuid = { workspace = true }
 thiserror = { workspace = true }
